# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "短链接生成与管理服务的 API 文档",
        "title": "短链接服务 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/captcha/send": {
            "post": {
                "description": "向指定邮箱或手机发送验证码，用于注册、登录、重置密码等场景",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证码"
                ],
                "summary": "发送验证码",
                "parameters": [
                    {
                        "description": "发送验证码请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SendCaptchaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SendCaptchaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "429": {
                        "description": "发送过于频繁",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/captcha/verify": {
            "post": {
                "description": "验证用户输入的验证码是否正确",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "验证码"
                ],
                "summary": "验证验证码",
                "parameters": [
                    {
                        "description": "验证请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VerifyCaptchaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "验证码错误或已过期",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.SendCaptchaRequest": {
            "type": "object",
            "required": [
                "account",
                "scene"
            ],
            "properties": {
                "account": {
                    "description": "手机号或邮箱",
                    "type": "string"
                },
                "scene": {
                    "description": "场景 (register/login/reset_pwd等)",
                    "type": "string"
                },
                "type": {
                    "description": "类型，默认为email",
                    "type": "string",
                    "enum": [
                        "sms",
                        "email"
                    ]
                }
            }
        },
        "dto.SendCaptchaResponse": {
            "type": "object",
            "properties": {
                "expire_second": {
                    "description": "验证码有效期（秒）",
                    "type": "integer"
                },
                "next_send_second": {
                    "description": "下次可发送的冷却时间（秒）",
                    "type": "integer"
                }
            }
        },
...
        "dto.VerifyCaptchaRequest": {
            "type": "object",
            "required": [
                "account",
                "captcha",
                "scene"
            ],
            "properties": {
                "account": {
                    "type": "string"
                },
                "captcha": {
                    "description": "验证码通常为6位",
                    "type": "string"
                },
                "scene": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}